20200219，实现回复功能以及异常处理规范化

+++ src/main/java/com/yzc/forum/controller/PublishController.java
@@ -1,7 +1,6 @@
 package com.yzc.forum.controller;
 
 import com.yzc.forum.dto.QuestionDTO;
-import com.yzc.forum.mapper.QuestionMapper;
 import com.yzc.forum.model.Question;
 import com.yzc.forum.model.User;
 import com.yzc.forum.service.QuestionService;
@@ -22,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,Model model){
+    public String edit(@PathVariable(name = "id") Long id,Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
         model.addAttribute("description",question.getDescription());
@@ -41,7 +40,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model){
